consistency - set all long fields to ints to match the data on the backend backend

+++ src/org/mythtv/client/ui/dvr/EpisodeFragment.java
@@ -189,7 +189,7 @@ public class EpisodeFragment extends AbstractMythFragment {
 		return super.onOptionsItemSelected( item );
 	}
 
-	public void loadEpisode( Long channelId, DateTime startTime ) {
+	public void loadEpisode( int channelId, DateTime startTime ) {
 		Log.v( TAG, "loadEpisode : enter" );
 
 		if( null == mRecordedDaoHelper ) {
